deprecate "multiple" option
