+++ addon-test-support/properties/has-class.js
@@ -98,7 +98,7 @@ export function hasClass(cssClass, selector, userOptions = {}) {
     isDescriptor: true,
 
     get(key) {
-      let options = assign({ pageObjectKey: key }, userOptions);
+      let options = assign({ pageObjectKey: key, __multiple__: true }, userOptions);
 
       let elements = options.multiple ? findMany(this, selector, options) : [findOne(this, selector, options)];
