+++ addon-test-support/properties/not-has-class.js
@@ -102,7 +102,7 @@ export function notHasClass(cssClass, selector, userOptions = {}) {
     isDescriptor: true,
 
     get(key) {
-      let options = assign({ pageObjectKey: key }, userOptions);
+      let options = assign({ pageObjectKey: key, __multiple__: true }, userOptions);
 
       let elements = options.multiple ? findMany(this, selector, options) : [findOne(this, selector, options)];
